# services/core-api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpportunityOS Core API
  version: 0.8.0
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /v1/sources:
    post:
      operationId: createSource
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses: { "201": { description: Product-neutral Source created } }
  /v1/sources/{id}/signals:
    parameters: [{ $ref: "#/components/parameters/ObjectID" }, { $ref: "#/components/parameters/IdempotencyKey" }]
    post:
      operationId: importSignal
      responses: { "201": { description: Signal imported or returned from fingerprint deduplication } }
  /v1/signals/{id}/promotions:
    parameters: [{ $ref: "#/components/parameters/ObjectID" }, { $ref: "#/components/parameters/IdempotencyKey" }]
    post:
      operationId: promoteSignal
      responses: { "201": { description: Opportunity and evidence created with explicit Signal lineage } }
...
  /v1/outcome-feedback:
    post:
      operationId: recordOutcomeFeedback
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "201": { description: Settled and reconciled order outcome recorded }
        "409": { description: Lineage, completion, settlement, or reconciliation prerequisite failed }
  /v1/adapter-identities:
    post:
      operationId: registerAdapterIdentity
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses: { "201": { description: Provider-bound Adapter key identity registered with an environment secret reference } }
  /v1/executions/{id}/workflow-runs:
    parameters: [{ $ref: "#/components/parameters/ObjectID" }, { $ref: "#/components/parameters/IdempotencyKey" }]
    post:
      operationId: startWorkflowRun
      responses: { "201": { description: Persistent WorkflowRun and retryable side-effect steps created } }
  /v1/workflow-steps/leases:
    post:
      operationId: leaseWorkflowStep
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses: { "200": { description: Due step leased to a matching Adapter identity } }
  /v1/adapter-results:
    post:
      operationId: ingestAdapterResult
      security: []
      parameters:
        - { in: header, name: X-Adapter-Key, required: true, schema: { type: string } }
        - { in: header, name: X-Adapter-Timestamp, required: true, schema: { type: string } }
        - { in: header, name: X-Adapter-Nonce, required: true, schema: { type: string, minLength: 16 } }
        - { in: header, name: X-Adapter-Signature, required: true, schema: { type: string } }
      responses:
        "201": { description: Signed result receipt persisted and applied under the owned workflow lease }
        "401": { description: Signature, timestamp, nonce, identity, or lease validation failed }
...
  /v1/operational-alerts/{id}/acknowledgements:
    parameters: [{ $ref: "#/components/parameters/ObjectID" }, { $ref: "#/components/parameters/IdempotencyKey" }]
    post:
      operationId: acknowledgeOperationalAlert
      responses: { "200": { description: Operational alert acknowledged } }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: opportunity_session
...
  parameters:
    ObjectID:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 1 }
security:
  - sessionCookie: []